Struct cbuf::CBuf

pub struct CBuf<'a, T: 'a> {
    // some fields omitted
}

Circular Buffer

Turns a slice into a Circular buffer with head and tail indexes.

Methods

impl<'a, T: Clone> CBuf<'a, T> where T: Clone

fn new(buf: &'a mut [T]) -> CBuf<T>

Create new CBuf

Length (not capacity) will be used to store elements in the circular buffer.

fn is_full(&self) -> bool

Is buffer full?

fn is_empty(&self) -> bool

Is buffer empty?

fn peek(&mut self) -> Option<T>

Peek next element from the CBuf without removing it

Returns None if buffer is empty.

fn peek_unchecked(&mut self) -> T

Peek next element from the CBuf without removing it

Makes the buffer misbehave if it's empty.

fn get(&mut self) -> Option<T>

Remove one element from the CBuf

Returns None if buffer is empty.

fn get_unchecked(&mut self) -> T

Remove one element from the CBuf

Makes the buffer misbehave if it's empty.

fn put(&mut self, val: T)

Add element the buffer

Ignores the write if buffer is full.

fn put_unchecked(&mut self, val: T)

Add element the buffer

Makes the buffer misbehave if it's full.